Removes unnecessary reader APIs
